requirements.txt
requests==2.31.0
Flask==3.0.3
lmdb==1.4.1
python-dotenv==1.0.1